requirements.txt
Django==6.0.1
Pillow==12.1.0
djangorestframework==3.15.2
django-filter==24.3
django-simple-history==3.7.0
django-import-export==4.1.1
openpyxl==3.1.5
flake8==7.1.1

